requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
flask-cors>=3.0.0
google-generativeai>=0.3.0
numpy>=1.21.0
opencv-python>=4.5.0
pandas>=1.3.0
Pillow>=8.0.0
pyautogui>=0.9.53
pygetwindow>=0.0.9
pyperclip>=1.8.2
pytesseract>=0.3.8
pywebview>=4.0.0
